requirements.txt
# === 核心框架 ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# === LLM API ===
openai>=1.10.0          # DeepSeek 兼容 OpenAI 接口
httpx>=0.26.0           # 异步 HTTP 客户端
dashscope>=1.24.6       # 阿里云百炼 SDK（Qwen-Plus, Qwen3-TTS, Qwen3-ASR）

# === 语音处理 ===
edge-tts>=6.1.9         # TTS 领读
openai-whisper>=20231117  # ASR 语音识别
sounddevice>=0.4.6      # 音频录制
soundfile>=0.12.1       # 音频文件读写

# === 发音评估 ===
jiwer>=3.0.3            # 词错误率计算
python-Levenshtein>=0.23.0  # 编辑距离

# === FSRS 间隔重复 ===
fsrs>=3.0.0             # 开源 FSRS 算法

# === 数据库 ===
aiosqlite>=0.19.0       # 异步 SQLite
sqlalchemy>=2.0.25      # ORM

# === Web 前端支持 ===
jinja2>=3.1.3           # 模板引擎
python-multipart>=0.0.9 # 文件上传
aiofiles>=23.2.1        # 异步文件操作
websockets>=12.0        # WebSocket 支持

# === 工具库 ===
python-dotenv>=1.0.0    # 环境变量管理
pyyaml>=6.0.1           # 配置文件
loguru>=0.7.2           # 日志
rich>=13.7.0            # 终端美化
